+++ RTNeural/conv1d/conv1d_eigen.h
@@ -11,8 +11,8 @@ template <typename T>
 class Conv1D : public Layer<T>
 {
 public:
-    Conv1D(size_t in_size, size_t out_size, size_t kernel_size, size_t dilation);
-    Conv1D(std::initializer_list<size_t> sizes);
+    Conv1D(int in_size, int out_size, int kernel_size, int dilation);
+    Conv1D(std::initializer_list<int> sizes);
     Conv1D(const Conv1D& other);
     Conv1D& operator=(const Conv1D& other);
     virtual ~Conv1D();
@@ -28,7 +28,7 @@ public:
         state.col(state_ptr) = inVec;
         state.col(state_ptr + state_size) = inVec;
 
-        for(size_t i = 0; i < Layer<T>::out_size; ++i)
+        for(int i = 0; i < Layer<T>::out_size; ++i)
             outVec(i, 0) = state.block(0, state_ptr, Layer<T>::in_size, state_size).cwiseProduct(kernelWeights[i]).sum();
 
         outVec = outVec + bias;
@@ -40,26 +40,26 @@ public:
     void setWeights(const std::vector<std::vector<std::vector<T>>>& weights);
     void setBias(const std::vector<T>& biasVals);
 
-    size_t getKernelSize() const noexcept { return kernel_size; }
-    size_t getDilationRate() const noexcept { return dilation_rate; }
+    int getKernelSize() const noexcept { return kernel_size; }
+    int getDilationRate() const noexcept { return dilation_rate; }
 
 private:
-    const size_t dilation_rate;
-    const size_t kernel_size;
-    const size_t state_size;
+    const int dilation_rate;
+    const int kernel_size;
+    const int state_size;
 
     std::vector<Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic>> kernelWeights;
     Eigen::Matrix<T, Eigen::Dynamic, 1> bias;
 
     Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic> state;
-    size_t state_ptr = 0;
+    int state_ptr = 0;
 
     Eigen::Matrix<T, Eigen::Dynamic, 1> inVec;
     Eigen::Matrix<T, Eigen::Dynamic, 1> outVec;
 };
 
 //====================================================
-template <typename T, size_t in_sizet, size_t out_sizet, size_t kernel_size, size_t dilation_rate>
+template <typename T, int in_sizet, int out_sizet, int kernel_size, int dilation_rate>
 class Conv1DT
 {
     using vec_type = Eigen::Matrix<T, out_sizet, 1>;
@@ -86,7 +86,7 @@ public:
         state.col(state_ptr) = ins;
         state.col(state_ptr + state_size) = ins;
 
-        for(size_t i = 0; i < out_size; ++i)
+        for(int i = 0; i < out_size; ++i)
             outs(i) = state.block(0, state_ptr, in_size, state_size).cwiseProduct(weights[i]).sum();
 
         outs = outs + bias;
@@ -97,8 +97,8 @@ public:
     void setWeights(const std::vector<std::vector<std::vector<T>>>& weights);
     void setBias(const std::vector<T>& biasVals);
 
-    constexpr size_t getKernelSize() const { return kernel_size; }
-    constexpr size_t getDilationRate() const { return dilation_rate; }
+    constexpr int getKernelSize() const { return kernel_size; }
+    constexpr int getDilationRate() const { return dilation_rate; }
 
     Eigen::Map<vec_type, Eigen::Aligned16> outs;
 
@@ -106,7 +106,7 @@ private:
     T outs_internal alignas(16)[out_size];
 
     state_type state;
-    size_t state_ptr = 0;
+    int state_ptr = 0;
 
     weights_type weights[out_size];
     vec_type bias;
