+++ RTNeural/conv1d/conv1d_eigen.tpp
@@ -4,14 +4,14 @@ namespace RTNeural
 {
 
 template <typename T>
-Conv1D<T>::Conv1D(size_t in_size, size_t out_size, size_t kernel_size, size_t dilation)
+Conv1D<T>::Conv1D(int in_size, int out_size, int kernel_size, int dilation)
     : Layer<T>(in_size, out_size)
     , dilation_rate(dilation)
     , kernel_size(kernel_size)
     , state_size(kernel_size * dilation)
 {
     kernelWeights.resize(out_size);
-    for(size_t i = 0; i < out_size; ++i)
+    for(int i = 0; i < out_size; ++i)
         kernelWeights[i] = Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic>::Zero(in_size, state_size);
 
     bias = Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic>::Zero(out_size, 1);
@@ -21,7 +21,7 @@ Conv1D<T>::Conv1D(size_t in_size, size_t out_size, size_t kernel_size, size_t di
 }
 
 template <typename T>
-Conv1D<T>::Conv1D(std::initializer_list<size_t> sizes)
+Conv1D<T>::Conv1D(std::initializer_list<int> sizes)
     : Conv1D<T>(*sizes.begin(), *(sizes.begin() + 1), *(sizes.begin() + 2), *(sizes.begin() + 3))
 {
 }
@@ -53,25 +53,25 @@ void Conv1D<T>::reset()
 template <typename T>
 void Conv1D<T>::setWeights(const std::vector<std::vector<std::vector<T>>>& weights)
 {
-    for(size_t i = 0; i < Layer<T>::out_size; ++i)
-        for(size_t k = 0; k < Layer<T>::in_size; ++k)
-            for(size_t j = 0; j < kernel_size; ++j)
+    for(int i = 0; i < Layer<T>::out_size; ++i)
+        for(int k = 0; k < Layer<T>::in_size; ++k)
+            for(int j = 0; j < kernel_size; ++j)
                 kernelWeights[i](k, j * dilation_rate) = weights[i][k][j];
 }
 
 template <typename T>
 void Conv1D<T>::setBias(const std::vector<T>& biasVals)
 {
-    for(size_t i = 0; i < Layer<T>::out_size; ++i)
+    for(int i = 0; i < Layer<T>::out_size; ++i)
         bias(i, 0) = biasVals[i];
 }
 
 //====================================================
-template <typename T, size_t in_sizet, size_t out_sizet, size_t kernel_size, size_t dilation_rate>
+template <typename T, int in_sizet, int out_sizet, int kernel_size, int dilation_rate>
 Conv1DT<T, in_sizet, out_sizet, kernel_size, dilation_rate>::Conv1DT()
     : outs(outs_internal)
 {
-    for(size_t k = 0; k < out_size; ++k)
+    for(int k = 0; k < out_size; ++k)
         weights[k] = weights_type::Zero();
 
     bias = vec_type::Zero();
@@ -79,26 +79,26 @@ Conv1DT<T, in_sizet, out_sizet, kernel_size, dilation_rate>::Conv1DT()
     reset();
 }
 
-template <typename T, size_t in_sizet, size_t out_sizet, size_t kernel_size, size_t dilation_rate>
+template <typename T, int in_sizet, int out_sizet, int kernel_size, int dilation_rate>
 void Conv1DT<T, in_sizet, out_sizet, kernel_size, dilation_rate>::reset()
 {
     state_ptr = 0;
     state = state_type::Zero();
 }
 
-template <typename T, size_t in_sizet, size_t out_sizet, size_t kernel_size, size_t dilation_rate>
+template <typename T, int in_sizet, int out_sizet, int kernel_size, int dilation_rate>
 void Conv1DT<T, in_sizet, out_sizet, kernel_size, dilation_rate>::setWeights(const std::vector<std::vector<std::vector<T>>>& ws)
 {
-    for(size_t i = 0; i < out_size; ++i)
-        for(size_t k = 0; k < in_size; ++k)
-            for(size_t j = 0; j < kernel_size; ++j)
+    for(int i = 0; i < out_size; ++i)
+        for(int k = 0; k < in_size; ++k)
+            for(int j = 0; j < kernel_size; ++j)
                 weights[i](k, j * dilation_rate) = ws[i][k][j];
 }
 
-template <typename T, size_t in_sizet, size_t out_sizet, size_t kernel_size, size_t dilation_rate>
+template <typename T, int in_sizet, int out_sizet, int kernel_size, int dilation_rate>
 void Conv1DT<T, in_sizet, out_sizet, kernel_size, dilation_rate>::setBias(const std::vector<T>& biasVals)
 {
-    for(size_t i = 0; i < out_size; ++i)
+    for(int i = 0; i < out_size; ++i)
         bias(i) = biasVals[i];
 }
