+++ RTNeural/gru/gru_accelerate.tpp
@@ -4,7 +4,7 @@ namespace RTNeural
 {
 
 template <typename T>
-GRULayer<T>::GRULayer(size_t in_size, size_t out_size)
+GRULayer<T>::GRULayer(int in_size, int out_size)
     : Layer<T>(in_size, out_size)
     , zWeights(in_size, out_size)
     , rWeights(in_size, out_size)
@@ -21,7 +21,7 @@ GRULayer<T>::GRULayer(size_t in_size, size_t out_size)
 }
 
 template <typename T>
-GRULayer<T>::GRULayer(std::initializer_list<size_t> sizes)
+GRULayer<T>::GRULayer(std::initializer_list<int> sizes)
     : GRULayer<T>(*sizes.begin(), *(sizes.begin() + 1))
 {
 }
@@ -51,7 +51,7 @@ GRULayer<T>::~GRULayer()
 }
 
 template <typename T>
-GRULayer<T>::WeightSet::WeightSet(size_t in_size, size_t out_size)
+GRULayer<T>::WeightSet::WeightSet(int in_size, int out_size)
     : out_size(out_size)
 {
     W = new T*[out_size];
@@ -59,7 +59,7 @@ GRULayer<T>::WeightSet::WeightSet(size_t in_size, size_t out_size)
     b[0] = new T[out_size];
     b[1] = new T[out_size];
 
-    for(size_t i = 0; i < out_size; ++i)
+    for(int i = 0; i < out_size; ++i)
     {
         W[i] = new T[in_size];
         U[i] = new T[out_size];
@@ -72,7 +72,7 @@ GRULayer<T>::WeightSet::~WeightSet()
     delete[] b[0];
     delete[] b[1];
 
-    for(size_t i = 0; i < out_size; ++i)
+    for(int i = 0; i < out_size; ++i)
     {
         delete[] W[i];
         delete[] U[i];
@@ -85,9 +85,9 @@ GRULayer<T>::WeightSet::~WeightSet()
 template <typename T>
 void GRULayer<T>::setWVals(const std::vector<std::vector<T>>& wVals)
 {
-    for(size_t i = 0; i < Layer<T>::in_size; ++i)
+    for(int i = 0; i < Layer<T>::in_size; ++i)
     {
-        for(size_t k = 0; k < Layer<T>::out_size; ++k)
+        for(int k = 0; k < Layer<T>::out_size; ++k)
         {
             zWeights.W[k][i] = wVals[i][k];
             rWeights.W[k][i] = wVals[i][k + Layer<T>::out_size];
@@ -99,9 +99,9 @@ void GRULayer<T>::setWVals(const std::vector<std::vector<T>>& wVals)
 template <typename T>
 void GRULayer<T>::setWVals(T** wVals)
 {
-    for(size_t i = 0; i < Layer<T>::in_size; ++i)
+    for(int i = 0; i < Layer<T>::in_size; ++i)
     {
-        for(size_t k = 0; k < Layer<T>::out_size; ++k)
+        for(int k = 0; k < Layer<T>::out_size; ++k)
         {
             zWeights.W[k][i] = wVals[i][k];
             rWeights.W[k][i] = wVals[i][k + Layer<T>::out_size];
@@ -113,9 +113,9 @@ void GRULayer<T>::setWVals(T** wVals)
 template <typename T>
 void GRULayer<T>::setUVals(const std::vector<std::vector<T>>& uVals)
 {
-    for(size_t i = 0; i < Layer<T>::out_size; ++i)
+    for(int i = 0; i < Layer<T>::out_size; ++i)
     {
-        for(size_t k = 0; k < Layer<T>::out_size; ++k)
+        for(int k = 0; k < Layer<T>::out_size; ++k)
         {
             zWeights.U[k][i] = uVals[i][k];
             rWeights.U[k][i] = uVals[i][k + Layer<T>::out_size];
@@ -127,9 +127,9 @@ void GRULayer<T>::setUVals(const std::vector<std::vector<T>>& uVals)
 template <typename T>
 void GRULayer<T>::setUVals(T** uVals)
 {
-    for(size_t i = 0; i < Layer<T>::out_size; ++i)
+    for(int i = 0; i < Layer<T>::out_size; ++i)
     {
-        for(size_t k = 0; k < Layer<T>::out_size; ++k)
+        for(int k = 0; k < Layer<T>::out_size; ++k)
         {
             zWeights.U[k][i] = uVals[i][k];
             rWeights.U[k][i] = uVals[i][k + Layer<T>::out_size];
@@ -141,9 +141,9 @@ void GRULayer<T>::setUVals(T** uVals)
 template <typename T>
 void GRULayer<T>::setBVals(const std::vector<std::vector<T>>& bVals)
 {
-    for(size_t i = 0; i < 2; ++i)
+    for(int i = 0; i < 2; ++i)
     {
-        for(size_t k = 0; k < Layer<T>::out_size; ++k)
+        for(int k = 0; k < Layer<T>::out_size; ++k)
         {
             zWeights.b[i][k] = bVals[i][k];
             rWeights.b[i][k] = bVals[i][k + Layer<T>::out_size];
@@ -155,9 +155,9 @@ void GRULayer<T>::setBVals(const std::vector<std::vector<T>>& bVals)
 template <typename T>
 void GRULayer<T>::setBVals(T** bVals)
 {
-    for(size_t i = 0; i < 2; ++i)
+    for(int i = 0; i < 2; ++i)
     {
-        for(size_t k = 0; k < Layer<T>::out_size; ++k)
+        for(int k = 0; k < Layer<T>::out_size; ++k)
         {
             zWeights.b[i][k] = bVals[i][k];
             rWeights.b[i][k] = bVals[i][k + Layer<T>::out_size];
@@ -167,7 +167,7 @@ void GRULayer<T>::setBVals(T** bVals)
 }
 
 template <typename T>
-T GRULayer<T>::getWVal(size_t i, size_t k) const noexcept
+T GRULayer<T>::getWVal(int i, int k) const noexcept
 {
     T** set = zWeights.W;
     if(k > 2 * Layer<T>::out_size)
@@ -185,7 +185,7 @@ T GRULayer<T>::getWVal(size_t i, size_t k) const noexcept
 }
 
 template <typename T>
-T GRULayer<T>::getUVal(size_t i, size_t k) const noexcept
+T GRULayer<T>::getUVal(int i, int k) const noexcept
 {
     T** set = zWeights.U;
     if(k > 2 * Layer<T>::out_size)
@@ -203,7 +203,7 @@ T GRULayer<T>::getUVal(size_t i, size_t k) const noexcept
 }
 
 template <typename T>
-T GRULayer<T>::getBVal(size_t i, size_t k) const noexcept
+T GRULayer<T>::getBVal(int i, int k) const noexcept
 {
     T** set = zWeights.b;
     if(k > 2 * Layer<T>::out_size)
